study(step4): 자동차 이름 추가
