+++ README.md
@@ -30,8 +30,8 @@
 ### 기능 목록
 
 - 자동차
-    - [ ] 이름을 가지고 있고, 5자를 초과하지 않는다.
-    - [ ] 움직인 위치 정보를 가지고 있다.
+    - [x] 이름을 가지고 있고, 5자를 초과하지 않는다.
+    - [x] 움직인 위치 정보를 가지고 있다.
 - 입력
     - [ ] 이름을 입력 받는다.
     - [ ] 입력 된 이름은 쉼표를 기준으로 여러 이름을 구분한다.
