+++ src/main/kotlin/racingcar/domain/model/CarRacer.kt
@@ -3,10 +3,14 @@ package racingcar.domain.model
 /**
  * 자동차 경주자
  */
-class CarRacer(position: Int = 0) {
+class CarRacer(val name: String, position: Int = 0) {
 
     var position: Int = position
         private set
+
+    init {
+        require(name.length <= 5) { "자동차 이름은 5자를 초과할 수 없습니다." }
+    }
 
     fun moveTo(randomCount: Int) {
         if (randomCount < 4) return
@@ -15,7 +19,10 @@ class CarRacer(position: Int = 0) {
 
     companion object {
         fun ofCarRacer(racer: CarRacer): CarRacer {
-            return CarRacer(racer.position)
+            return CarRacer(
+                name = racer.name,
+                position = racer.position
+            )
         }
     }
 }
